WIP: Update MRI projects for Aluminium
This bumps versions to the following:
- odlparent-7.0.1
- yangtools-5.0.1
- mdsal-6.0.0
- controller-2.0.0

Implements the List->Map migration due to the MD-SAL 6 upgrade and
partially the migration to Uint{8,16,32} types.

WIP: updated lisp-proto, dsbackend, implementation, southbound, neutron.
SFTs and integrationtests are missing.

+++ mappingservice/lisp-proto/src/main/java/org/opendaylight/lispflowmapping/lisp/serializer/EidRecordSerializer.java
@@ -12,6 +12,7 @@ import org.opendaylight.lispflowmapping.lisp.serializer.address.LispAddressSeria
 import org.opendaylight.lispflowmapping.lisp.serializer.address.LispAddressSerializerContext;
 import org.opendaylight.lispflowmapping.lisp.util.ByteUtil;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.lfm.lisp.proto.rev151105.eid.container.Eid;
+import org.opendaylight.yangtools.yang.common.Uint8;
 
 public final class EidRecordSerializer {
     private static final EidRecordSerializer INSTANCE = new EidRecordSerializer();
@@ -26,7 +27,7 @@ public final class EidRecordSerializer {
 
     public Eid deserialize(ByteBuffer requestBuffer) {
         /* byte reserved = */requestBuffer.get();
-        short maskLength = (short) (ByteUtil.getUnsignedByte(requestBuffer));
+        Uint8 maskLength = Uint8.valueOf(ByteUtil.getUnsignedByte(requestBuffer));
         LispAddressSerializerContext ctx = new LispAddressSerializerContext(maskLength);
         return LispAddressSerializer.getInstance().deserializeEid(requestBuffer, ctx);
     }
